first commit for the new repo.
the "foundation" code is complete aka the fundemntals for the NN and genetic algorithim code.

Project was ripped from one folder to another so work will be needed to get netbeans to accept it.

bugtesting for tomorrow to figure out where the issue is, and then reorganization for OOP for the rest of the week.